chore(stackdriver): Refactored Stackdriver into gcp_service base class
This is functionally equivalent but factors out functionality in such a way
as to be able to reuse it when introducing ServiceControl support.

The besides moving code around, the main refactoring is how time series data
is constructed.
